pyproject.toml
[tool.poetry]
name = "git-pusher"
version = "0.1.2"
description = "Push files in multiple git repos in one go"
authors = ["John Stilia <stilia.johny@gmail.com>"]
license = "GPL-3"
keywords = ["git", "multi-push"]
readme = "README.md"
include = [
    "LICENSE.txt",
    "README.md",
]
homepage = "https://stiliajohny.github.io/python-git-pusher "
repository = "https://github.com/stiliajohny/python-git-pusher.git "


[tool.poetry.dependencies]
python = "^3.10"
pyfiglet = "^0.8.post1"
GitPython = "^3.1.26"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
git-pusher = 'git_pusher:main'
